examples script for workforce scheduling problem.
