+++ discrete_optimization/workforce/allocation/utils.py
