+++ examples/workforce/scheduling/run_dashboard.py
@@ -1,0 +1,13 @@
+from discrete_optimization.generic_tools.dashboard import Dashboard
+from discrete_optimization.generic_tools.study import Hdf5Database
+
+study_name = "scheduling-study-0"
+
+# retrieve data
+with Hdf5Database(f"{study_name}.h5") as database:
+    results = database.load_results()
+
+
+# launch dashboard with this data
+app = Dashboard(results=results)
+app.run()
