+++ examples/workforce/scheduling/run_relaxed.py
@@ -1,0 +1,44 @@
+#  Copyright (c) 2025 AIRBUS and its affiliates.
+#  This source code is licensed under the MIT license found in the
+#  LICENSE file in the root directory of this source tree.
+import logging
+
+from discrete_optimization.generic_tools.cp_tools import ParametersCp
+from discrete_optimization.workforce.scheduling.parser import (
+    get_data_available,
+    parse_json_to_problem,
+)
+from discrete_optimization.workforce.scheduling.solvers.cpsat import (
+    CPSatAllocSchedulingSolver, ObjectivesEnum
+)
+from discrete_optimization.workforce.scheduling.solvers.cpsat_relaxed import (
+    CPSatAllocSchedulingSolverCumulative
+)
+from discrete_optimization.generic_tools.callbacks.early_stoppers import (
+    ObjectiveGapStopper,
+)
+from discrete_optimization.generic_tools.callbacks.loggers import NbIterationTracker
+from discrete_optimization.generic_tools.callbacks.stats_retrievers import (
+    BasicStatsCallback,
+    StatsWithBoundsCallback,
+)
+import os
+logging.basicConfig(level=logging.INFO)
+
+def run_cumulative():
+    instances = [p for p in get_data_available()]
+    problem = parse_json_to_problem(instances[1])
+    solver = CPSatAllocSchedulingSolverCumulative(problem)
+    solver.init_model(objectives=[ObjectivesEnum.NB_TEAMS],
+                      adding_redundant_cumulative=True)
+    res = solver.solve_two_step(callbacks=[ObjectiveGapStopper(0, 0),
+                                           BasicStatsCallback()],
+                                parameters_cp=ParametersCp.default_cpsat(),
+                                time_limit=10,
+                                ortools_cpsat_solver_kwargs={"log_search_progress": True})
+    sol = res[-1][0]
+    print(problem.satisfy(sol), problem.evaluate(sol))
+
+
+if __name__ == "__main__":
+    run_cumulative()
